Striker

+++ module/purpose/Striker/src/Striker.cpp
@@ -69,7 +69,7 @@ namespace module::purpose {
         on<Provide<NormalStriker>, When<Phase, std::equal_to, Phase::READY>>().then([this] {
             // If we are stable, walk to the ready field position
             emit<Task>(std::make_unique<WalkToFieldPosition>(
-                Eigen::Vector3f(cfg.ready_position.x(), cfg.ready_position.y(), 0),
+                Eigen::Vector3d(cfg.ready_position.x(), cfg.ready_position.y(), 0),
                 cfg.ready_position.z()));
         });
